+++ CHANGELOG.md
@@ -8,6 +8,7 @@ See [GitHub releases](https://github.com/jenkinsci/jira-plugin/releases)
 ### Unreleased
 Release date:  _March 2, 2020_
 * Changed all references of JIRA to Jira per [Atlassian branding updates](https://community.atlassian.com/t5/Feedback-Forum-articles/A-new-look-for-Atlassian/ba-p/638077)
+* JiraCreateIssueNotifier: regard all statuses in "done" category as finished
 
 ### 3.0.11 
 Release date:  _Nov 21, 2019_